Updated to latest versions
I uploaded the latest code for the new and beta builds

+++ Source/ImprovedFactions/ImprovedFactions_new/src/main/java/io/github/toberocat/core/commands/admin/AdminDisbandSubCommand.java
@@ -6,25 +6,31 @@ import io.github.toberocat.core.utility.command.SubCommandSettings;
 import io.github.toberocat.core.utility.factions.Faction;
 import io.github.toberocat.core.utility.language.LangMessage;
 import io.github.toberocat.core.utility.language.Language;
+import io.github.toberocat.core.utility.language.Parseable;
 import org.bukkit.entity.Player;
 
 import java.util.List;
 
 public class AdminDisbandSubCommand extends SubCommand  {
     public AdminDisbandSubCommand() {
-        super("disband", LangMessage.COMMAND_ADMIN_DISBAND_DESCRIPTION, false);
+        super("disband", "admin.disband", LangMessage.COMMAND_ADMIN_DISBAND_DESCRIPTION, false);
     }
 
     @Override
     public SubCommandSettings getSettings() {
-        return super.getSettings().setArgLength(1);
+        return super.getSettings().setArgLength(1).setUseWhenFrozen(true);
     }
 
     @Override
     protected void CommandExecute(Player player, String[] args) {
         Faction faction = FactionUtility.getFactionByRegistry(args[0]);
+        if (faction == null) {
+            Language.sendRawMessage("&cCan't find given faction", player);
+            return;
+        }
         faction.delete();
-        Language.sendMessage(LangMessage.COMMAND_ADMIN_DISBAND_SUCCESS, player);
+        Language.sendMessage(LangMessage.COMMAND_ADMIN_DISBAND_SUCCESS, player,
+                new Parseable("{faction_display}", faction.getDisplayName()));
     }
 
     @Override
